refactor: Create helpers folder in desktop/electron
