+++ ui/desktop/electron-app/src/helpers/spawn-promise.js
